chore: simplify encrypting payload when handling DecryptionError
Previously, we used the 'local' lockbox instance when re-encrypting
the payload to log. This adds complexity and potential confusion
when decrypting the payload, since we'd need to reconstruct an
instance with the same context (table, attribute, whether or not
to use a the org key) when attempting to diagnose the issue.

This removes that layer of complexity, and allows one to just
create a lockbox instance that uses the `master_key` when
decrypting the payload.

Of course, once we get the payload decrypted, we may still run into
a version of the problem I describe above anyway, but at least
this can remove one potential layer of confusion.